fix issues, add README.MD, now checking on existing other yaml files

+++ internal/repository/profileDelete.go
@@ -6,6 +6,11 @@ import (
 )
 
 func (y *YamlRepository) ProfileDelete(name string) error {
+	profile, _ := LoadProfile(name)
+	if profile == nil {
+		return fmt.Errorf("profile %s not found", name)
+	}
+
 	err := os.Remove(fmt.Sprintf("%s.yaml", name))
 	if err != nil {
 		return err
